Fix existing specs
