+++ codegen/projections/white_label/lib/white_label/stubs.rb
@@ -194,7 +194,7 @@ module WhiteLabel
 
       def self.stub(http_resp, stub:)
         data = {}
-        stub_io = stub[:output] === String ? StringIO.new(stub[:output]) : stub[:output]
+        stub_io = String === stub[:output] ? StringIO.new(stub[:output]) : stub[:output]
         IO.copy_stream(stub_io, http_resp.body)
       end
     end
